requirements.txt
aiobotocore
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
annotated-doc
annotated-types
anyio
appnope
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
babel
beautifulsoup4
bleach
boto3
botocore
certifi
cffi
charset-normalizer
choreographer
click
comm
contourpy
cryptography
cycler
debugpy
decorator
defusedxml
docstring_parser
dotenv
executing
faiss-cpu
fastapi
fastjsonschema
filelock
fonttools
fqdn
frozenlist
fsspec
greenlet
h11
h5py
httpcore
httptools
httpx
httpx-sse
hypothesis
idna
importlib_metadata
iniconfig
IProgress
ipykernel
ipython
ipython_pygments_lexers
ipywidgets
isoduration
jedi
Jinja2
jmespath
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kaleido
kiwisolver
lark
llvmlite
logistro
MarkupSafe
matplotlib
matplotlib-inline
mcp
mistune
mpmath
multidict
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numba
numpy
opentelemetry-api
opentelemetry-instrumentation
opentelemetry-instrumentation-threading
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
packaging
pandas
pandoc
pandocfilters
parso
pexpect
pillow
platformdirs
playwright
plotly
pluggy
plumbum
ply
prometheus_client
prompt_toolkit
propcache
psutil
ptyprocess
pure_eval
pycparser
pydantic
pydantic-settings
pydantic_core
pyee
Pygments
PyJWT
pynndescent
pyparsing
pytest
pytest-timeout
python-dateutil
python-dotenv
python-json-logger
python-multipart
PyYAML
pyzmq
referencing
requests
rfc3339-validator
rfc3986-validator
rfc3987-syntax
rpds-py
s3fs
s3transfer
scikit-learn
scipy
Send2Trash
setuptools
simplejson
six
sortedcontainers
soupsieve
sse-starlette
stack-data
starlette
strands-agents
sympy
terminado
threadpoolctl
tinycss2
torch
tornado
tqdm
traitlets
typing-inspection
typing_extensions
tzdata
umap-learn
uri-template
urllib3
uvicorn
uvloop
watchdog
watchfiles
wcwidth
webcolors
webencodings
websocket-client
websockets
widgetsnbextension
wrapt
yarl
zipp
